NewExtend, and Concat are added
The newExtend adds a way to make a 2D slice where the inner slices
are made with double the capacity. This is useful for the cases where
2D slices are needed that can be expanded without a need for reallocation
(until the capacity is reached).

Also the Concant function and its test were added, to combine two 2D slices
by appending each row of the two together.